Add Discord -> SMF thread creation syncing
Mirrorer reacts with a globe emoticon when it has synced. The thread and post then appear in the SMF forum under Mirrorer's account.
A text tag is just added to the start of all posts naming the user who posted the message. Don't yet know enough about php templates to slot it into the sidebar but it should be easy enough implementation, the poster_name field of the messages database contains the name of the Discord user.