Add Discord -> SMF message deletion
Also fixed a bug where topic titles were undefined due to a missing parameter to getSubjectLine().
If the deleted message is the first in a thread, its database entry is just updated to say the message has been deleted. This maintains the thread.
If the deleted message is a normal body message, it is just deleted in both SMF and mirror databases.
If the deleted message is the last in a thread, its topic's database entry is updated to say its preceeding message is the last in the thread.